UIRS-62 Refactor Configurations API: decouple data-specific aspects
- sorting
- preparing values for mutations
