+++ src/data/Configurations/index.js
@@ -1,0 +1,2 @@
+export * from './queries';
+export * from './mutations';
