+++ src/data/index.js
@@ -1,2 +1,3 @@
+export * as Configurations from './Configurations';
 export * as Locations from './Locations';
 export * as AccessionTable from './AccessionTable';
